fix: move flattenSuggestions out of the use-server module
'use server' files may only export async functions; the sync helper
now lives in lib/ai/nodeSuggestions.ts. Restores the production
build broken by the previous commit.
