+++ actions/ai-node-edit.ts
@@ -7,6 +7,7 @@ import {
   generatedNodeSuggestionsSchema,
   type GeneratedTreeNode,
 } from "@/lib/ai/generatedTree";
+import { flattenSuggestions } from "@/lib/ai/nodeSuggestions";
 import { AIConfigurationError, getAnthropicModel } from "@/lib/anthropic";
 import { AIMindmap } from "@/types/AI";
 
@@ -14,45 +15,6 @@ const NODE_EDIT_INSTRUCTIONS = `You extend one selected branch of a mindmap.
 Return only useful new child nodes beneath the selected node. Create two to four
 levels of concise information, with no more than three children per node. Do not
 repeat existing branch content. Descriptions and links are optional.`;
-
-/** Converts a generated nested suggestion tree into the legacy flat contract. */
-function flattenSuggestions(
-  nodes: GeneratedTreeNode[],
-  activeNode: AIMindmap
-): AIMindmap[] {
-  let sequence = 0;
-  const flattened: AIMindmap[] = [];
-
-  /** Assigns internal relationship ids that are replaced by canvas ids later. */
-  function visit(node: GeneratedTreeNode): string {
-    sequence += 1;
-    const nodeId = `suggestion-${sequence}`;
-    const childrenNodes = node.children.map(visit);
-
-    flattened.push({
-      nodeId,
-      title: node.title,
-      description: node.description ?? null,
-      link: node.link ?? null,
-      childrenNodes,
-    });
-
-    return nodeId;
-  }
-
-  const childrenNodes = nodes.map(visit);
-
-  return [
-    {
-      ...activeNode,
-      nodeId: activeNode.nodeId,
-      // Suggestions extend the selected branch; they must not erase siblings
-      // already present in the canvas's flat tree.
-      childrenNodes: [...(activeNode.childrenNodes ?? []), ...childrenNodes],
-    },
-    ...flattened,
-  ];
-}
 
 /**
  * Generates nested children for NodeAiEdit while preserving its existing flat
@@ -104,4 +66,4 @@ Current branch: ${JSON.stringify(currentBranch)}`,
   }
 }
 
-export { editAIMindmap, flattenSuggestions };
+export { editAIMindmap };
